fix(aot): el hash de ELF salia distinto segun donde se compilara Vesta
Mismo fallo de fondo que el umbral del JIT, y este es peor porque el sintoma no
aparece en quien lo produce.

El hash SysV de ELF -- el que indexa la tabla `.hash` / `DT_HASH` de los
simbolos dinamicos -- estaba escrito sobre `unsigned long`:

    unsigned long h = 0, g;
    h = (h << 4) + (unsigned char)*name++;
    g = h & 0xf0000000UL;

Esta forma de escribirlo se apoya en que `h << 4` SE SALGA por arriba, y eso
solo pasa donde `unsigned long` mide 32 bits.  En Windows mide 32 y funciona;
en Linux mide 64, los bits por encima de 32 se acumulan y la mascara
`0xf0000000` no los limpia nunca, asi que el hash sale OTRO.

Y lo lee el enlazador dinamico del sistema, que calcula el de 32 bits del
estandar.  Un Vesta construido en Linux emitia una tabla `.hash` que ld.so no
sabe recorrer, y el sintoma no es un error del emisor: es un simbolo que "no
existe" al cargar el binario.  Lo usan los dos emisores, el de x86-64 y el de
x86-32.

Se arregla con anchura FIJA (`uint32_t`), que es lo que el formato define y lo
unico que vale igual en toda plataforma.

Encontrado barriendo el arbol por el patron del umbral del JIT.  El resto de
`unsigned long` que quedan se miraron y se dejan: son indices de TLS y el
`GetExceptionCode()` de Win32, donde ese es el tipo de la API.

+++ src/aot/common/aot_emit_internal.h
@@ -190,6 +190,10 @@ extern int g_aot_dbg_n;
 
 /* Hash SysV de ELF (para .hash / DT_HASH).  Usado por los dynexec 64/32
  * (aun en los emisores por-arch); definido en aot_emit_common.c. */
-unsigned long aot_elf_hash(const char *name);
+/* 32 bits FIJOS: el hash esta definido asi y se apoya en salirse por arriba.
+ * Con `unsigned long` salia distinto en Linux (64 bits) que en Windows (32), y
+ * la tabla `.hash` que emitiamos no la sabia recorrer el enlazador dinamico.
+ * Ver la nota completa en la definicion. */
+uint32_t aot_elf_hash(const char *name);
 
 #endif /* AOT_COMMON_AOT_EMIT_INTERNAL_H */
